+++ src/com/openscg/pgstudio/client/providers/ItemListProvider.java
@@ -6,5 +6,5 @@ package com.openscg.pgstudio.client.providers;
 import com.openscg.pgstudio.client.PgStudio.ITEM_TYPE;
 
 public interface ItemListProvider extends ListProvider {
-	public void setItem(int schema, int item, ITEM_TYPE type);	
+	public void setItem(int schema, long item, ITEM_TYPE type);	
 }
